+++ modules/tinymce/src/core/main/ts/textpatterns/keyboard/Keyboard.ts
@@ -44,13 +44,7 @@ const setup = (editor: Editor): void => {
     if (!isSelectionCollapsed && hasPatterns) {
       // Passing through just the part of PatternSet that is relevant for inline
       // patterns.
-      KeyHandler.handleInlineKey(
-        editor,
-        {
-          inlinePatterns: patternSet.inlinePatterns,
-          dynamicPatternsLookup: patternSet.dynamicPatternsLookup
-        }
-      );
+      KeyHandler.handleInlineKey(editor, patternSet);
     }
   };
